Fancier/requirements.txt
anyascii==0.3.0
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coverage==6.3
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0
django-allauth==0.49.0
django-cors-headers==3.10.1
django-filter==21.1
django-modelcluster==5.2
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docutils==0.18.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filelock==3.4.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
l18n==2021.3
Markdown==3.3.6
my==1.3.0
mysqlclient==2.1.0
nose2==0.10.0
numpy==1.22.3
oauthlib==3.2.0
openpyxl==3.0.9
pandas==1.4.1
Pillow==9.0.1
platformdirs==2.4.1
playsound==1.3.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyreadline==2.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.1
sqlclient==21.4.24
sqlparse==0.4.2
tablib==3.2.0
telepath==0.2
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.0
wagtail==2.16.1
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
